ref(onboarding): Split bottle onboarding docs
