+++ static/app/gettingStartedDocs/python/bottle/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './bottle';
+import docs from '.';
 
 describe('bottle onboarding docs', () => {
   it('renders doc correctly', () => {
